[YT-22] sidebar 반응형 로직 개발 / 3가지 상태
Main, search 페이지에서 데스크탑, 태블릿, 모바일에 따라 3가지 상태의 사이드바를 가짐반응형에 맞춰서 로직 개발
