+++ src/pages/Main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import MainContainer from '../views/main/containers/MainContainer';
-import SidebarContainer from '../views/shared/containers/SidebarContainer';
+import SidebarContainer from '../views/shared/containers/Sidebar/SidebarContainer';
 
 const Main = () => (
   <Container>
