+++ src/pages/Search.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import SearchContainer from '../views/search/containers/SearchContainer';
-import SidebarContainer from '../views/shared/containers/SidebarContainer';
+import SidebarContainer from '../views/shared/containers/Sidebar/SidebarContainer';
 
 const Search = () => (
   <Container>
